requirements.txt
# Core LangChain
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.1
langsmith>=0.0.11  # For tracing and observability

# Document Loading & Processing
pypdf2>=3.0.0      # For basic PDF text extraction
unstructured[all]>=0.10.0  # For more robust PDF parsing (tables, elements)
python-magic>=0.4.24 # Required by 'unstructured' for filetype detection

# Vector Database & Embeddings
chromadb>=0.4.0    # Our vector store
openai>=1.0.0       # For embeddings and LLMs

# Caching & Memory
redis>=4.5.0        # Our cache and session store

# Utilities
python-dotenv>=1.0.0 # For loading .env files
tiktoken>=0.5.0      # For counting tokens (used by LangChain)

# UI (We'll use Streamlit for the MVP)
streamlit>=1.28.0

pre-commit>=3.0.0   # For running linters before commits

# Unit testing
pytest>=9.0.0

# For running Jupyter Notebooks
ipykernel==7.1.0